fix(ts): Follow same definition architecture as official 'vue' and 'vuex' typedefs
And also same codestyle (double quotes, semi-colons for imports)
