+++ distributions/nuxt/types/vue.d.ts
@@ -1,0 +1,28 @@
+/**
+ * Extends interfaces in Vue.js
+ */
+
+import Vue, { ComponentOptions } from "vue";
+import { Route } from "vue-router";
+import { MetaInfo } from "vue-meta";
+import { Context, Transition } from "./index";
+
+declare module "vue/types/options" {
+  interface ComponentOptions<V extends Vue> {
+    asyncData?: (ctx: Context) => object;
+    fetch?: (ctx: Context) => Promise<void> | void;
+    head?: MetaInfo | (() => MetaInfo);
+    layout?: string | ((ctx: Context) => string);
+    middleware?: string | string[];
+    scrollToTop?: boolean;
+    transition?: string | Transition | ((to: Route, from: Route) => string);
+    validate?: (ctx: Context) => Promise<boolean> | boolean;
+    watchQuery?: boolean | string[];
+  }
+}
+
+declare module "vue/types/vue" {
+  interface Vue {
+    $nuxt: any; //TODO: Provide better type definition
+  }
+}
